Implement a latency counter for multiplayer mode Intention is to use this for lag compensation
Basic implementation with both sides sending
a ping SYN packet every 50 turns and responding
with a ping ACK. Transmission times are recorded
along the way.

Not sure if this is a good way to do it, but I
got headtied thinking about doing this without
a ping packet, so I just went ahead and made
one.

Recording more times than are
probably useful.