Creative Mode for developers
I added a creative mode for developers. It's only changeable in the
source code -> Player.java -> boolean creative(set this to true).

Things added in this mode
----------------------------------
* free buildings/rails.
* Instant-kill on everything you can imagine.
* Aimbot v. 1.0

I hope you guys enjoy this and it would be cool if you would be able to
give me some feedback. I'm going to try include some free-upgrading in
the next release, but for now just would like to hear some feedback!