Refix for issue #100, problem introduced by #369 Loss of random floor tiles after #369
I didn't check when changing tile images to make
sure I was only changing shadowed floor tiles.

This caused a common path to revert all floor tiles
without shadows to be the same image.

Updated now so that when a floor tile is changed
from shadowed to non-shadowed, it gets a random
floor tile image.