Implement variable width font Also small test font in 3 shades
This is intended to help implement translation
of the cool new shop item tooltips which use
very small fonts.

To make the font look halfway decent, the chars
need to be variable width.

The test font is a copy of the font in the tooltip
image, except with my rubbish improvisation for
the characters that were missing. Feel free
to fix them.

Also, neatened up the Font class while I was there.