Fix a problem where game start freezes in debug Eclipse debugging freezes due to exceptions
Rather than doing a null check in Mob() we were
just catching NullPointerExceptions which caused
the debugger to pause at the "failing" line.

I think it was creating a bunch of Mob objects to
show the map selection page, but the difficulty
had not been set yet.

Changed to a null check and debugger is happy again.