Fix marker and highlighting of Buildings in MP Interaction with enemy buildings not allowed Previously there were a number of issues with highlighting and markers. This patch changes the functionality to the following:
* You may only pick up or upgrade buildings you own OR
  buildings owned by the Neutral team. This no longer gives
  a message, it just doesn't work any more (see next point).

* An interaction marker (the pulsing square reticle) only
  appears over a building you can interact with (see above).
  And it only appears on your game client.

* Tool tips only appear on ShopItems you own and only on your
  own game client.

The code in Player#pickup() that shows the message
"You can not pick that up!" when you try and pick up an enemy
building should no longer be executed, but I have not removed
it.

Hopefully the fact that the marker is no longer shown should
make it obvious enough that you cannot interact with the
Building.

That is all.