Fix issue #89: player ported to top-left of map
Cause was placing an active mouse cursor at the direct
center of the character (causing the aim vector to have
no direction) AND firing.

You get warning output even if you aren't firing, but
you suffer no ill effect. If you are firing, then
you get ported to the top-left of the map.

Fix is not to update aim vector if the mouse is at
the center of the character.