Fixed problem with Fonts in DownloadScreen.
Even if this fix was fast, this is a signal of a worst problems comming:
1- we have 2 static AbstractScreen objects (MojamComponent & MojamStartup).
2- Fonts are instanciated statically. If some class try to use Fonts before
MojamComponent.screen is set, NullPointerException will come.

For now, im giving priority to MojamComponent.screen one. MojamStartup will
use it if set, or create and set. MojamComponent will overwrite regardless
if its set or not.

To avoid exception, an easy solution would be create a static void
CreateFonts(AbstractScreen) and call it from *Screen constructors.

I think it is a big matter to do it w/o asking first.